+++ changelog.md
@@ -3,6 +3,12 @@ This file contains all the notable changes done to the Ballerina MQTT package th
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Fixed
+
+- [Allow onComplete method in the service declaration](https://github.com/ballerina-platform/ballerina-library/issues/7272)
+
 ## [1.1.1] - 2024-07-17
 
 ### Fixed